Make cinaps compatible with ocaml 4.04, 4.05 and 4.06
The only thing preventing it from being compatible with those
older compilers was the use of `StringLabels.split_on_char` which was
introduced in 4.05.
I tried building it locally and it worked well, I'm not sure why it
had a 4.07 lower bound in the first place so I might very well be
missing something!

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>